refactor(localized-link): changes according CR
