+++ libs/shared/src/index.ts
@@ -1,4 +1,5 @@
 export * from "./components";
+export * from "./constants";
 export * from "./hooks";
 export * from "./types";
 export * from "./utils";
