+++ libs/shared/src/types/app-routes.ts
@@ -16,12 +16,3 @@ export enum PagePath {
   PresentationForExternalUser = "external-user-presentation",
   Dashboard = "dashboard"
 }
-
-export const PAGE_PATHS: Record<PagePath, AppRoute | AppRoute[]> = {
-  [PagePath.Dashboard]: AppRoute.Dashboard,
-  [PagePath.Home]: AppRoute.Home,
-  [PagePath.Presentation]: AppRoute.Presentation,
-  [PagePath.AddPresentation]: [AppRoute.Presentation, AppRoute.AddPresentation],
-  [PagePath.EditPresentation]: [AppRoute.Presentation, AppRoute.EditPresentation],
-  [PagePath.PresentationForExternalUser]: [AppRoute.Presentation, AppRoute.PresentationForExternalUser]
-};
