+++ libs/shared/src/utils/create-path.ts
@@ -5,5 +5,6 @@ import { AppRoute, SupportedLanguage } from "../types";
 export const createPath = (path: AppRoute | AppRoute[], params?: Params, language?: string) => {
   const lang = language?.length && language !== SupportedLanguage.En ? `?lang=${language}` : "";
   const basePath = path instanceof Array ? path.join("/") : path;
+
   return `${generatePath(`${basePath}`, params)}${lang}`;
 };
